move test case to mangle-subst.cpp
